stop passing jobsPanel through functions; use constant instead
